Implement a ServerNetworkHandler, based on the ClientNetworkHandler Update network server to remove all game logic from it (not yet reimplemented elsewhere) Clean all warnings, especially deprecated strncpy and APIENTRY

+++ src/client/MainHandler.cpp
@@ -5,12 +5,12 @@
 #include <iostream>
 #include <iomanip>
 // GL
-#include "GL/glew.h"
-#include "GLFW/glfw3.h"
-#include "glm/ext.hpp"
 #include "MainHandler.hh"
 #include "RenderHandler.hh"
 #include "KeyMapper.hh"
+#include "GL/glew.h"
+#include "GLFW/glfw3.h"
+#include "glm/ext.hpp"
 
 namespace bl {
 	MainHandler::MainHandler() :
@@ -157,7 +157,7 @@ namespace bl {
 			double x,
 			double y
 	) {
-		this->m_activeBrowser.lock()->mouseScroll(x, y);
+		this->m_activeBrowser.lock()->mouseScroll(static_cast<int>(x), static_cast<int>(y));
 	}
 
 	void MainHandler::onCursorEnter(
